Promote incremental subpackage to flat layout, restore pre-merge architecture
Reverses the incremental-subsystem outcome of merging main's PR #317 while
keeping its functional improvements. Files at flat module paths match the
pre-merge `incremental_*` naming; the generator and scope_planner return to
the pre-merge `derive_patch_scopes` / `apply_patch_scopes` flow returning
Path. Subpackage `diagram_analysis/incremental/` deleted.

Layout:
- `incremental/{delta+updater}.py` -> `incremental_updater.py` (combined,
  matching pre-merge file shape).
- `incremental/{tracer+trace_planner}.py` -> `incremental_tracer.py`
  (combined, matching pre-merge file shape).
- `incremental/models.py` -> `incremental_models.py`.
- `incremental/{payload,pipeline}.py` -> `incremental_{payload,pipeline}.py`.
- `incremental/semantic_diff.py` -> `static_analyzer/semantic_diff.py` (its
  conceptual home; pre-merge had it split there as cosmetic_diff +
  method_fingerprint + tree_sitter_parsers).

API restored to pre-merge:
- `IncrementalUpdater(analysis, symbol_resolver, repo_dir, ...)` with
  `compute_delta(added_files, modified_files, deleted_files, changes)`.
- `run_trace(..., parsed_diff=None, callbacks=None)`.
- `TraceResult.impacted_methods` (not `all_impacted_methods`).
- `MethodResolver` uses `graph_query.resolve_method_node` (matches
  pre-merge tracer + existing test mocks of `get_reference` /
  `get_loose_reference`).

Functional improvements kept (strictly additive over pre-merge):
- `TraceConfig` budget bundling.
- Retries with backoff in the trace LLM call.
- Anthropic prompt caching.
- `classify_scope` for component mapping.
- `prune_empty_components`, `drop_deltas_for_pruned_components`.
- `COSMETIC_ONLY` stop reason + reporting.
- Modern `repo_utils.change_detector.ChangeSet` parsing layer.

User's tweaks preserved:
- `agent_llm` (not `parsing_llm`) plumbed through `apply_patch_scopes` /
  `patch_analysis_scope` for incremental patching.
- In-place `apply_method_delta` (rather than tuple-returning `apply_delta`).
- `skip_cache=self.force_full_analysis` for the injected-analyzer path.

Deletions:
- `agents/analysis_patcher.py` (duplicate; canonical is the pre-merge
  `diagram_analysis/analysis_patcher.py`).
- Tests of post-merge-only generator semantics (TestGenerateAnalysisIncrementalFailedPatches,
  test_incremental_generator_fallback_sets_top_level_full_required_flag) that
  the pre-merge architecture cannot express.
- `plan.md`, `workflow.md` (planning docs from main's PR).
- `pyproject.toml`: drop `diagram_analysis.incremental` declaration.
