+++ static_analyzer/constants.py
@@ -49,7 +49,7 @@ assert set(LANGUAGE_EXTENSIONS) == set(
 
 # Flattened reverse lookup: extension -> language. Used by the diff boundary
 # (``repo_utils/diff_parser.py``) to filter non-source changes and by
-# ``diagram_analysis.incremental.semantic_diff`` / ``incremental.tracer`` to
+# ``static_analyzer.semantic_diff`` / ``diagram_analysis.incremental_tracer`` to
 # route files to the right parser. Derived from ``LANGUAGE_EXTENSIONS`` so
 # adding a language in one place updates both.
 SOURCE_EXTENSION_TO_LANGUAGE: dict[str, Language] = {
