+++ tests/diagram_analysis/test_incremental_tracer_plan.py
@@ -13,12 +13,12 @@ from textwrap import dedent
 import pytest
 
 from agents.change_status import ChangeStatus
-from diagram_analysis.incremental.trace_planner import (
+from diagram_analysis.incremental_tracer import (
     ChangeGroup,
     _collapse_fallback_groups,
     build_trace_plan,
 )
-from diagram_analysis.incremental.delta import FileDelta, IncrementalDelta, MethodChange
+from diagram_analysis.incremental_updater import FileDelta, IncrementalDelta, MethodChange
 from repo_utils.change_detector import ChangeSet
 from static_analyzer.constants import NodeType
 from static_analyzer.graph import CallGraph
@@ -101,7 +101,7 @@ def test_cosmetic_only_modification_is_skipped_from_plan():
             cfgs={},
             repo_dir=repo,
             base_ref=base_ref,
-            change_set=ChangeSet(base_ref=base_ref, target_ref=""),
+            parsed_diff=ChangeSet(base_ref=base_ref, target_ref=""),
         )
         # Cosmetic-only file should contribute no groups
         assert plan.groups == []
@@ -195,7 +195,7 @@ def test_plan_builds_groups_for_modified_method_with_callers():
             cfgs={"python": cfg},
             repo_dir=repo,
             base_ref=base_ref,
-            change_set=ChangeSet(base_ref=base_ref, target_ref=""),
+            parsed_diff=ChangeSet(base_ref=base_ref, target_ref=""),
         )
         # Helper was modified (with a real body change), has an upstream caller -> should NOT be fast-path
         assert plan.fast_path_impacted_methods == []
@@ -252,7 +252,7 @@ def test_plan_fast_path_used_for_leaf_modification_with_no_callers():
             cfgs={"python": cfg},
             repo_dir=repo,
             base_ref=base_ref,
-            change_set=ChangeSet(base_ref=base_ref, target_ref=""),
+            parsed_diff=ChangeSet(base_ref=base_ref, target_ref=""),
         )
         # No callers, signature unchanged, body-only change -> fast path
         assert plan.fast_path_impacted_methods == ["src.utils.isolated"]
